contrived passing test suite to highlight coverage issues
